% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSIBMA.R
\name{MOSIBMA}
\alias{MOSIBMA}
\title{MOSIBMA}
\usage{
MOSIBMA(
  balance = TRUE,
  full_design = NULL,
  factor_level,
  unit,
  particle_number = 10,
  particle_increase = 10,
  SIB_time = 3,
  G_MA_list,
  q_GB = 1,
  q_LB = 1,
  q_new = 1,
  t = 10,
  t_increase = 10,
  structure_matrix = NULL,
  structure_factor = NULL,
  initial_design = NULL,
  iteration = 10
)
}
\arguments{
\item{balance}{If set to \code{TRUE}, balanced designs are considered and
the argument \code{full_design} is not required.
If the user considers unbalanced designs, the argument should set to \code{TRUE}.}

\item{full_design}{A data.frame stands for the complete factorial design.
Required only when \code{balance} is set to \code{FALSE}.}

\item{factor_level}{A list where each element is a numeric vector specifying
levels of a factor.}

\item{unit}{An integer representing the number of experimental units in a design.}

\item{particle_number}{An integer indicating the initial particle number.}

\item{particle_increase}{An integer indicating the number of particles increased sequentially.}

\item{SIB_time}{An integer indicating the number of MOSIBMA processes that will run simultaneously.
Should larger than 1.}

\item{G_MA_list}{A list where each element is a list of matrices standing for the orthogonal
projection matrices corresponding to a minimum aberration criterion.}

\item{q_GB}{An integer indicating the number of columns of a design that
should be mixed with the corresponding columns of a design in the global best particle.}

\item{q_LB}{An integer indicating the number of columns of a design that
should be mixed with the corresponding columns of a design in the local best particle.}

\item{q_new}{An integer indicating the number of columns of a design that
should be mixed with the corresponding columns of a random design.}

\item{t}{An integer indicating the initial iteration number.}

\item{t_increase}{An integer indicating the increased iteration number sequentially.}

\item{structure_matrix}{A list where each element is an incidence matrix.
Required only when groups of units need to have the same value of some factors, e.g., split-plot
or strip-plot design, and so on.}

\item{structure_factor}{A list where each element is a numeric vector specifying
which factors should have the same value in terms of an incidence matrix.}

\item{initial_design}{A list of data.frames representing designs that are promising to be minimum
aberration designs.}

\item{iteration}{An integer indicating the total iteration number of MOSIBMA.}
}
\value{
A list containing two elements:
\describe{
\item{GMA_design_candidate}{A list of candidate designs.}
\item{GMA_design.info}{A list consists \code{G_MA_index} and \code{G_MA_WL}.
\code{G_MA_index} represents the index of minimum aberration designs in \code{GMA_design_candidate}.
\code{G_MA_WL} represents the wordlength patterns of the minimum aberration designs.
The order of \code{G_MA_index} and \code{G_MA_WL} follows the order of \code{G_MA_list}.}
}
}
\description{
Use swarm intelligence-based algorithm to search for designs that satisfy the
corresponding minimum aberration criterion.
}
